pillow/requirements.txt
requests
Pillow
numpy
flask
google-cloud-storage
